refactor: cleaned the code from comments
cleaned the code from comments
