+++ my-app/src/utils/audio.js
@@ -1,7 +1,5 @@
-import blockMovement from '../audios/block_movement.mp3';
-
 const blockMovementPlayAudio = (volume = 0.2) => {
-    const sound = new Audio(blockMovement);
+    const sound = new Audio("./audios/block_movement.mp3");
     sound.volume = volume;
     return sound.play();
 }
